create Components folder, move files, and import useState & useEffect to App
