+++ src/App.js
@@ -1,6 +1,7 @@
 import './App.css';
-import VehicleList from './VehicleList';
-import TrafficLight from './TrafficLight';
+import { useState, useEffect } from 'react';
+import VehicleList from './Components/VehicleList';
+import TrafficLight from './Components/TrafficLight';
 
 function App() {
   // track the following state with a few useState hooks:
@@ -8,7 +9,8 @@ function App() {
   // lizardSize should be a number that starts out as 10
   // alienSize should be a number that starts out as 10
   // traffic is complicated. It should be an array of strings that starts out as ['car', 'truck']
-
+ 
+  
   return (
     <div className="App">
       <div className="fight">
