+++ src/Components/Vehicle.js
@@ -1,6 +1,6 @@
 export default function Vehicle() {
   return <div className="vehicle">
-    {/* the Vehcile component has a prop called vehicle. Check to see what string is in this vehicle prop */}
+    {/* the Vehicle component has a prop called vehicle. Check to see what string is in this vehicle prop */}
     {/* if the vehicle props is a car, render a '🚗' emoji */}
     {/* if the vehicle props is a truck, render a '🚚' emoji */}
     {/* if the vehicle props is a bus, render a '🚌' emoji */}
